Add new TextMonitor classes.
